PLI improvements
The procedure should allow the selection of multiple videos for space-time diagram extraction. The selected reference line / circle will be previewed and selected using the last video only but the reference line / circle will be applied to all videos selected.

If two images are selected for interval scaling, then they should not be merged for scaling. In this case, clicking engage should enable the user to add an arbitrary space between the tow images, and click 'Refresh' until the interval lines overlap with the first image.